+++ packages/yourloops/test/integration/pages/daily/daily-view-for-HCP.spec.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, Diabeloop
+ * Copyright (c) 2022-2023, Diabeloop
  *
  * All rights reserved.
  *
@@ -25,7 +25,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { screen } from '@testing-library/react'
+import { screen, waitFor } from '@testing-library/react'
 import { mockAuth0Hook } from '../../mock/auth0.hook.mock'
 import { mockTeamAPI } from '../../mock/team.api.mock'
 import { mockDataAPI } from '../../mock/data.api.mock'
@@ -56,7 +56,10 @@ describe('Daily view for HCP', () => {
 
   it('should render correct layout', async () => {
     mockDataAPI()
-    renderPage(`/patient/${unmonitoredPatientId}/daily`)
+    const router = renderPage(`/patient/${unmonitoredPatientId}/daily`)
+    await waitFor(() => {
+      expect(router.state.location.pathname).toEqual(`/patient/${unmonitoredPatientId}/daily`)
+    })
 
     expect(await screen.findByTestId('patient-nav-bar', {}, { timeout: 3000 })).toBeVisible()
     checkPatientNavBarAsHCP()